+++ src/api-token-provider.ts
@@ -1,6 +1,6 @@
 import { IAuthProvider } from "./iauth-provider";
-import jwt from "jsonwebtoken";
-import crypto from "crypto";
+import * as jwt from "jsonwebtoken";
+import * as crypto from "crypto";
 import { v4 as uuid } from "uuid";
 
 export class ApiTokenProvider implements IAuthProvider {