+++ CHANGELOG.md
@@ -2,11 +2,30 @@
 
 All notable changes to this project will be documented in this file. Dates are displayed in UTC.
 
+#### [v2.5.3](https://github.com/fireblocks/fireblocks-sdk-js/compare/v2.4.2...v2.5.3)
+
+> 8 December 2022
+
+- Fireblocks SDK DOS vulnerability - bump query-string version [`#126`](https://github.com/fireblocks/fireblocks-sdk-js/pull/126)
+
+#### [v2.4.2](https://github.com/fireblocks/fireblocks-sdk-js/compare/v2.4.1...v2.4.2)
+
+> 23 November 2022
+
+- update CreateTransactionResponse [`#122`](https://github.com/fireblocks/fireblocks-sdk-js/pull/122)
+
+#### [v2.4.1](https://github.com/fireblocks/fireblocks-sdk-js/compare/v2.4.0...v2.4.1)
+
+> 21 November 2022
+
+- add proxy support for axios instance [`#116`](https://github.com/fireblocks/fireblocks-sdk-js/pull/116)
+
 #### [v2.4.0](https://github.com/fireblocks/fireblocks-sdk-js/compare/v2.3.2...v2.4.0)
 
 > 26 October 2022
 
 - Add network related functionality [`#101`](https://github.com/fireblocks/fireblocks-sdk-js/pull/101)
+- Added method to get all audits for selected time period [`#114`](https://github.com/fireblocks/fireblocks-sdk-js/pull/114)
 
 #### [v2.3.2](https://github.com/fireblocks/fireblocks-sdk-js/compare/v2.3.1...v2.3.2)